+++ src/Yarn/Lock/Types.hs
@@ -19,8 +19,7 @@ import qualified Data.List.NonEmpty as NE
 --
 -- Attention: Might be changed to a newtype in a future release.
 type Lockfile = MKM.MKMap PackageKey Package
--- TODO
--- newtype Lockfile = Lockfile (MKM.MKMap PackageKey Package)
+-- TODO newtype Lockfile = Lockfile (MKM.MKMap PackageKey Package)
 
 -- | Proxy type for our MKMap intermediate key
 lockfileIkProxy :: Proxy Int
@@ -34,8 +33,8 @@ data PackageKey = PackageKey
   -- sometimes a npm semver, sometimes an arbitrary string
   } deriving (Show, Eq, Ord)
 
--- The name of a package. They can be scoped, see
--- <https://docs.npmjs.com/misc/scope> for an explanation.
+-- | The name of a package. They can be scoped, see
+-- | <https://docs.npmjs.com/misc/scope> for an explanation.
 data PackageKeyName
   = SimplePackageKey Text
   -- ^ just a package name
@@ -43,6 +42,7 @@ data PackageKeyName
   -- ^ a scope and a package name (e.g. @types/foobar)
   deriving (Show, Eq, Ord)
 
+-- | Try to parse a string into a package key name (scoped or not).
 parsePackageKeyName :: Text -> Maybe PackageKeyName
 parsePackageKeyName n = case T.stripPrefix "@" n of
   Nothing -> Just $ SimplePackageKey n